feat: Add forwardEmail, replyEmail, replyAllEmail, and sendEmail methods
This commit adds new methods to the EmailConnectorGraphMsService class. The forwardEmail method allows forwarding an email to a specified recipient with an optional comment. The replyEmail method allows replying to an email with a comment. The replyAllEmail method allows replying to all recipients of an email with a comment. The sendEmail method allows sending an email with a specified message.